refactor: 完成项目全面重构，全面使用BigDecimal替代double处理金额
这是一次大规模的项目重构，主要变更包括：
1.  数据库连接配置新增SSL和时区选项，优化JDBC连接参数
2.  全量替换double为BigDecimal进行金额计算，解决浮点精度问题
3.  重构交易、日志、转账等数据模型，适配BigDecimal类型
4.  重写货币格式化逻辑，修复千分位和精度处理问题
5.  移除过时的ThreadUtils工具类，改用Bukkit原生异步调度
6.  新增支付冷却功能，优化命令权限控制
7.  重构全局统计和缓存管理，新增缓存保存选项
8.  升级数据库版本系统，添加表结构升级脚本，优化字段精度
9.  修复离线转账、排行榜等功能的适配问题
10. 优化日志系统，使用新的时间格式化API
11. 新增/优化多个命令的功能，包括income命令支持查看其他玩家收入
12. 重构事件系统，同时支持原始double和新的BigDecimal类型参数

+++ src/main/java/com/oolonghoo/wooeco/database/dao/LogDAO.java
@@ -3,6 +3,7 @@ package com.oolonghoo.wooeco.database.dao;
 import com.oolonghoo.wooeco.database.DatabaseManager;
 import com.oolonghoo.wooeco.model.EconomyLog;
 
+import java.math.BigDecimal;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -31,9 +32,9 @@ public class LogDAO {
             stmt.setString(1, log.getUuid().toString());
             stmt.setString(2, log.getPlayerName());
             stmt.setString(3, log.getAction());
-            stmt.setDouble(4, log.getAmount());
-            stmt.setDouble(5, log.getBalanceBefore());
-            stmt.setDouble(6, log.getBalanceAfter());
+            stmt.setBigDecimal(4, log.getAmount());
+            stmt.setBigDecimal(5, log.getBalanceBefore());
+            stmt.setBigDecimal(6, log.getBalanceAfter());
             stmt.setString(7, log.getOperator());
             stmt.setString(8, log.getOperatorName());
             stmt.setString(9, log.getReason());
@@ -59,9 +60,9 @@ public class LogDAO {
                     UUID.fromString(rs.getString("uuid")),
                     rs.getString("player_name"),
                     rs.getString("action"),
-                    rs.getDouble("amount"),
-                    rs.getDouble("balance_before"),
-                    rs.getDouble("balance_after"),
+                    rs.getBigDecimal("amount"),
+                    rs.getBigDecimal("balance_before"),
+                    rs.getBigDecimal("balance_after"),
                     rs.getString("operator"),
                     rs.getString("operator_name"),
                     rs.getString("reason"),
@@ -89,9 +90,9 @@ public class LogDAO {
                     UUID.fromString(rs.getString("uuid")),
                     rs.getString("player_name"),
                     rs.getString("action"),
-                    rs.getDouble("amount"),
-                    rs.getDouble("balance_before"),
-                    rs.getDouble("balance_after"),
+                    rs.getBigDecimal("amount"),
+                    rs.getBigDecimal("balance_before"),
+                    rs.getBigDecimal("balance_after"),
                     rs.getString("operator"),
                     rs.getString("operator_name"),
                     rs.getString("reason"),
